Reorganize bundle (wip)
