+++ src/Module/ModuleOffers.php
@@ -18,12 +18,17 @@ namespace WEM\OffersBundle\Module;
 use Contao\Config;
 use Contao\ContentModel;
 use Contao\FrontendTemplate;
+use Contao\Input;
 use Contao\Module;
+use Contao\PageModel;
 use Contao\RequestToken;
-use Contao\StringUtil;
 use Contao\System;
+use Contao\Validator;
+use WEM\OffersBundle\Model\Alert;
+use WEM\OffersBundle\Model\OfferFeed;
 use WEM\OffersBundle\Model\OfferFeedAttribute;
 use WEM\OffersBundle\Model\Offer;
+use WEM\UtilsBundle\Classes\StringUtil;
 
 /**
  * Common functions for job offers modules.
@@ -32,6 +37,145 @@ use WEM\OffersBundle\Model\Offer;
  */
 abstract class ModuleOffers extends Module
 {
+    protected function catchAjaxRequests()
+    {
+        if (Input::post('TL_AJAX') && (int) $this->id === (int) Input::post('module')) {
+            try {
+                switch (Input::post('action')) {
+                    case 'seeDetails':
+                        if (!Input::post('offer')) {
+                            throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['argumentMissing'], 'offer'));
+                        }
+                        $objItem = Offer::findByPk(Input::post('offer'));
+
+                        $this->offer_template = 'offer_details';
+                        echo \Haste\Util\InsertTag::replaceRecursively($this->parseOffer($objItem));
+                        exit;
+                    break;
+
+                    case 'apply':
+                        if (!Input::post('offer')) {
+                            throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['argumentMissing'], 'offer'));
+                        }
+
+                        // Put the offer in session
+                        $objSession->set('wem_offer', Input::post('offer'));
+
+                        echo \Haste\Util\InsertTag::replaceRecursively($this->getApplicationForm(Input::post('offer')));
+                        exit;
+                    break;
+
+                    case 'subscribe':
+                        // Check if we have a valid email
+                        if (!Input::post('email') || !Validator::isEmail(Input::post('email'))) {
+                            throw new \Exception($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['invalidEmail']);
+                        }
+
+                        // Check if we have conditions
+                        $arrConditions = [];
+                        if (Input::post('conditions')) {
+                            foreach (Input::post('conditions') as $c => $v) {
+                                $arrConditions[$c] = $v;
+                            }
+                        }
+
+                        // Check if we already have an existing alert with this email and this conditions
+                        if (0 < Alert::countItems(
+                            ['email' => Input::post('email'), 'feed' => $this->offer_feed, 'conditions' => $arrConditions, 'active' => 1]
+                        )) {
+                            throw new \Exception($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['alertAlreadyExists']);
+                        }
+
+                        // The alert might be inactive, so instead of delete it
+                        // and create a new alert, try to retrieve an existing but disable one
+                        $objAlert = Alert::findItems(
+                            ['email' => Input::post('email'), 'feed' => $this->offer_feed, 'conditions' => $arrConditions, 'active' => 0],
+                            1
+                        );
+
+                        if (!$objAlert) {
+                            $objAlert = new Alert();
+                            $objAlert->createdAt = time();
+                        }
+
+                        $objAlert->tstamp = time();
+                        $objAlert->lastJob = time();
+                        $objAlert->activatedAt = 0;
+                        $objAlert->email = Input::post('email');
+                        $objAlert->frequency = Input::post('frequency') ?: 'daily'; // @todo -> add default frequency as setting
+                        $objAlert->token = StringUtil::generateToken(); // @todo -> add code system to confirm requests as alternatives to links/token
+                        $objAlert->feed = $this->offer_feed; // @todo -> build a multi feed alert
+                        $objAlert->moduleOffersAlert = $this->id;
+                        $objAlert->language = $GLOBALS['TL_LANGUAGE'];
+                        $objAlert->save();
+
+                        if (!empty($arrConditions)) {
+                            foreach ($arrConditions as $c => $v) {
+                                $objAlertCondition = new AlertCondition();
+                                $objAlertCondition->tstamp = time();
+                                $objAlertCondition->createdAt = time();
+                                $objAlertCondition->pid = $objAlert->id;
+                                $objAlertCondition->field = $c;
+                                $objAlertCondition->value = $v;
+                                $objAlertCondition->save();
+                            }
+                        }
+
+                        // Build and send a notification
+                        $arrTokens = $this->getNotificationTokens($objAlert);
+                        $objNotification = Notification::findByPk($this->offer_ncSubscribe);
+                        $objNotification->send($arrTokens);
+
+                        // Write the response
+                        $arrResponse = [
+                            'status' => 'success',
+                            'msg' => $GLOBALS['TL_LANG']['WEM']['OFFERS']['MSG']['alertCreated'],
+                        ];
+                    break;
+
+                    case 'unsubscribe':
+                        // Check if we have a valid email
+                        if (!Input::post('email') || !Validator::isEmail(Input::post('email'))) {
+                            throw new \Exception($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['invalidEmail']);
+                        }
+
+                        $objAlert = Alert::findItems(['email' => Input::post('email'), 'feed' => $this->offer_feed], 1);
+
+                        // Check if the alert exists or if the alert is already active
+                        if (!$objAlert) {
+                            throw new \Exception($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['alertDoesNotExists']);
+                        }
+
+                        // Generate a token for this request
+                        $objAlert->token = StringUtil::generateToken(); // @todo -> add code system to confirm requests as alternatives to links/token
+                        $objAlert->save();
+
+                        // Check if the alert was not activated
+                        $arrTokens = $this->getNotificationTokens($objAlert);
+                        $objNotification = Notification::findByPk($this->offer_ncUnsubscribe);
+                        $objNotification->send($arrTokens);
+
+                        // Write the response
+                        $arrResponse = [
+                            'status' => 'success',
+                            'msg' => $GLOBALS['TL_LANG']['WEM']['OFFERS']['MSG']['requestSent'],
+                        ];
+                    break;
+
+                    default:
+                        throw new \Exception(sprintf($GLOBALS['TL_LANG']['WEM']['OFFERS']['ERROR']['unknownRequest'], Input::post('action')));
+                }
+            } catch (\Exception $e) {
+                $arrResponse = ['status' => 'error', 'msg' => $e->getResponse(), 'trace' => $e->getTrace()];
+            }
+
+            // Add Request Token to JSON answer and return
+            $arrResponse['rt'] = RequestToken::get();
+            echo json_encode($arrResponse);
+            exit;
+        }
+    }
+
     /**
      * Parse one or more items and return them as array.
      *
@@ -197,6 +341,42 @@ abstract class ModuleOffers extends Module
         $objTemplate->form = $strForm;
 
         return $objTemplate->parse();
+    }
+
+    /**
+     * Build Notification Tokens.
+     *
+     * @param Alert $objAlert
+     *
+     * @return array
+     */
+    protected function getNotificationTokens($objAlert)
+    {
+        $arrTokens = [];
+
+        $objFeed = OfferFeed::findByPk($objAlert->feed);
+        foreach ($objFeed->row() as $strKey => $varValue) {
+            $arrTokens['feed_'.$strKey] = $varValue;
+        }
+
+        foreach ($objAlert->row() as $strKey => $varValue) {
+            $arrTokens['subscription_'.$strKey] = $varValue;
+        }
+
+        if ($this->offer_pageSubscribe && $objSubscribePage = PageModel::findByPk($this->offer_pageSubscribe)) {
+            $arrTokens['link_subscribe'] = $objSubscribePage->getAbsoluteUrl().'?wem_action=subscribe&token='.$objAlert->token;
+        }
+
+        if ($this->offer_pageUnsubscribe && $objSubscribePage = PageModel::findByPk($this->offer_pageUnsubscribe)) {
+            $arrTokens['link_unsubscribe'] = $objSubscribePage->getAbsoluteUrl().'?wem_action=unsubscribe';
+            $arrTokens['link_unsubscribeConfirm'] = $objSubscribePage->getAbsoluteUrl().'?wem_action=unsubscribe&token='.$objAlert->token;
+        }
+
+        $arrTokens['recipient_email'] = $objAlert->email;
+
+        $arrTokens['admin_email'] = $GLOBALS['TL_ADMIN_EMAIL'];
+
+        return $arrTokens;
     }
 
     /**
